Non Shared Changed And feedback
